Added namespaces
Started a list of namespaces at the end of the document. They are a mix of project specific, project related, and standard namespaces. They are currently alphabetized by prefix but do not need to be. It may make sense to break them up into smaller related groupings at some point.